Refactor OpenAPI builder structure and consolidate docs
- Rename FromRoutes to Builder::Spec as main builder class
- Move helper modules to builder/concerns/ directory
- Consolidate documentation into single docs/openapi_3.md file
- Update architecture diagram to reflect new structure
