+++ lib/grape-swagger/openapi/builder/concerns/responses.rb
