+++ lib/grape-swagger/openapi/builder/concerns/schemas.rb
